requirements.txt
pytest>=8.0.0
pytest-xdist>=3.5.0
pytest-timeout>=2.2.0
pytest-rerunfailures>=13.0
pytest-html==4.1.1
pytest-json-report==1.5.0
schemathesis>=3.31.0
requests>=2.31.0
responses>=0.24.1
requests-mock>=1.11.0
faker==22.6.0
PyJWT==2.8.0
jwt>=1.0.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv==1.0.0
beautifulsoup4==4.12.3
jmespath==1.0.1
allure-pytest==2.13.2
urllib3==2.1.0
aiohttp>=3.9.1
